Fix long navigation items overlapping the top logo
If the navigation items contain long text (e.g. German, Italian,
Romanian versions) the navigation items sometimes overlap the logo at
the top of the site. This makes the site look somewhat broken. The
cause of this issue is mainly due to the centering of the text. This is
fixed by aligning the text to the left and adding a padding which
corresponds the width of the logo (161px). Additionally the beginning
padding of ul elements is removed and the spacing between the
navigation items are reduced to use the space more efficiently.